+++ package.json
@@ -36,6 +36,7 @@
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-icons": "^5.5.0",
+    "react-material-ui-carousel": "^3.4.2",
     "react-select": "^5.10.2",
     "styled-components": "^6.3.11",
     "uuid": "^13.0.0",
